fix zowe/zowe-cli#748 - exit 1 when Imperative.init fails
