+++ package.json
@@ -65,7 +65,7 @@
     "audit:public": "npm audit --registry https://registry.npmjs.org/"
   },
   "dependencies": {
-    "@zowe/imperative": "4.7.3",
+    "@zowe/imperative": "4.7.4",
     "@zowe/perf-timing": "1.0.7",
     "get-stdin": "7.0.0",
     "js-yaml": "3.13.1",
